Adding database

+++ app/src/main/java/com/example/myweatherapp/detailActivity/WeatherDetailActivity.kt
@@ -3,9 +3,9 @@ package com.example.myweatherapp.detailActivity
 import android.annotation.SuppressLint
 import android.content.Context
 import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.TextView
+import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModelProvider
 import com.example.myweatherapp.R
@@ -15,7 +15,7 @@ class WeatherDetailActivity : AppCompatActivity() {
 
     companion object {
         const val INTENT_ID_NAME = "WeatherId"
-        fun getIntent(context: Context, id: Int): Intent {
+        fun getIntent(context: Context, id: Long): Intent {
             val intent = Intent(context, WeatherDetailActivity::class.java)
             intent.putExtra(INTENT_ID_NAME, id)
             return intent
@@ -38,19 +38,19 @@ class WeatherDetailActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_weather_detail)
-        val id = intent.getIntExtra(INTENT_ID_NAME, 0)
+        val id = intent.getLongExtra(INTENT_ID_NAME, 0)
         viewModel = ViewModelProvider(this)[WeatherDetailViewModel::class.java]
         weather = viewModel.getWeather(id)
         weather.observe(this, {
             it?.let {
-            tvDayAndDate.text = it.getDayAndDate()
-            tvDayNightTemp.text = it.getDayNightTemp()
-            tvStatus.text = it.getWeatherDescription()
-            tvClouds.text = it.getCloudsValue()
-            tvHumidity.text = it.getHumidityValue()
-            tvPrecipitation.text = it.getPrecipitation()
-            tvWind.text = it.getWind()
-            tvPressure.text = it.getPressures()
+                tvDayAndDate.text = it.getDayAndDate()
+                tvDayNightTemp.text = it.getDayNightTemp()
+                tvStatus.text = it.getWeatherDescription()
+                tvClouds.text = it.getCloudsValue()
+                tvHumidity.text = it.getHumidityValue()
+                tvPrecipitation.text = it.getPrecipitation()
+                tvWind.text = it.getWind()
+                tvPressure.text = it.getPressures()
             }
         })
